+++ src/main/java/com/example/demo/DataProcessor/Menu.java
@@ -1,0 +1,14 @@
+package com.example.demo.DataProcessor;
+
+public class Menu {
+    public static void printMenu() {
+        String expectedOutput = "Меню:\n" +
+                "1 - Вывести запись по id\n" +
+                "2 - Найти записи по части наименования\n" +
+                "3 - выход\n" +
+                "4 - сохранить в json\n" +
+                "5 - добавить значение\n" +
+                "Выберите пункт меню: ";
+        System.out.println(expectedOutput);
+    }
+}
