+++ src/main/java/com/example/demo/Model/Record.java
@@ -1,0 +1,17 @@
+package com.example.demo.Model;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+
+public class Record {
+    private int id;
+    private String name;
+    private String description;
+    private String link;
+
+}
