--- src/main/java/org/example/Record.java
@@ -1,29 +1,0 @@
-package org.example;
-public class Record {
-    private int id;
-    private String name;
-    private String description;
-    private String link;
-    public Record(int id, String name, String description, String link) {
-        this.id = id;
-        this.name = name;
-        this.description = description;
-        this.link = link;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public String getLink() {
-        return link;
-    }
-}
