+++ src/main/resources/application.properties
@@ -1,0 +1,2 @@
+input.data.path=./src/main/resources/data.json
+input.test.path=./src/main/resources/
