--- src/main/resources/outputFilePath.json
@@ -1,1 +1,0 @@
-{"2":{"id":2,"name":"weqe","description":"23232","link":"32323"},"3":{"id":3,"name":"ere","description":"23232","link":"32323"},"23":{"id":23,"name":"we","description":"wew","link":"wew"}}
