--- src/main/resources/werwrwe.json
@@ -1,1 +1,0 @@
-[{"id":2,"name":"weqe","description":"23232","link":"32323"},{"id":434,"name":"ewrew","description":"343","link":"erw"},{"id":3,"name":"ere","description":"23232","link":"32323"},{"id":454,"name":"srea","description":"3fdas","link":"34"}]
